edit product fucntionality completed
