+++ frontend/src/App.js
@@ -5,9 +5,11 @@ import SignUp from "./pages/SignUp"
 import NewProduct from "./pages/NewProduct"
 import AccountPage from "./pages/AccountPage";
 import ShowProduct from './pages/ShowProduct';
+import ShowProductEdit from './pages/ShowProductEdit';
 import { Routes, Route } from 'react-router-dom';
 import { useState, useEffect } from 'react';
 import { getToken } from './utils/api';
+
 
 
 
@@ -43,6 +45,7 @@ function App() {
         <Routes>
           <Route path="/" element={<Home/>} />
           <Route path="product/:id" element={<ShowProduct currentUser={currentUser}/>} />
+          <Route path="/productedit/:id" element={<ShowProductEdit />} />
           <Route path="/login" element={<LogIn setIsLoggedIn={setIsLoggedIn} />} />
           <Route path="/signup" element={<SignUp setIsLoggedIn={setIsLoggedIn}/>} />
           <Route path="/newproduct" element={<NewProduct/>} />
