+++ frontend/src/pages/ShowProductEdit/index.js
@@ -1,28 +1,62 @@
-import { useEffect, useState } from "react";
-import { updateOneProduct } from "../../utils/api";
+import { useEffect, useState } from "react"
+import { useParams } from "react-router-dom"
+import EditProduct from "../../components/EditProduct"
+import { deleteOneProduct, updateOneProduct, showOneProduct } from "../../utils/api"
+import { useNavigate } from "react-router-dom"
 import { Form, Button } from 'react-bootstrap';
 
-function EditProduct(props) {
+const ShowProductEdit = () => {
+    const navigate = useNavigate()
 
+    const [showProductData, setShowProductData] = useState({title: '', description: '', image: '', location: '', price: ''})
+    const [formShow, setFormShow] = useState(false)
     
-    const [editedState, setEditedState] = useState(props.showProductData)
+    
+    const {id} = useParams()
+    useEffect(() => {
+        showOneProduct(id).then(data => {setShowProductData(data)})
+    }, [id])
+    
+    
+    const [editedState, setEditedState] = useState(showProductData)
 
     const handleChange = (event) => {
         setEditedState({ ...editedState, [event.target.id]: event.target.value })
     }
 
-    const handleSubmit = async (event) => {
+    const handleSubmit = (event) => {
         event.preventDefault()
-        updateOneProduct(props.showProductData._id).then(data => {
-            
-            props.setShowProductData(data)
+        updateOneProduct(showProductData._id, editedState).then(data => {
+            setShowProductData(data)
         })
         
     }
+    
 
+    // Deletes then navigates to home page
+    const deleteProduct = () => {
+        const itemId = showProductData._id
+        deleteOneProduct(itemId).then(alert("You deleted this product"))
+        navigate('/')
+    }
 
+   
     return(
-            <Form className="edit-form" onSubmit={handleSubmit}>
+        <>
+        <div>
+        <img src={showProductData.image} />
+        <h1>{showProductData.title}</h1>
+        <p>{showProductData.description}</p>
+        
+
+
+        </div>
+       
+        <div>
+
+        <button onClick={deleteProduct}>Delete</button> 
+        </div>
+        <Form className="edit-form" onSubmit={handleSubmit}>
                 <Form.Group className="m-0">
                     <Form.Label>Title</Form.Label>
                     <Form.Control
@@ -83,6 +117,9 @@ function EditProduct(props) {
                 >Submit Edit
                 </Button>
             </Form>
-        )
-    }
-export default EditProduct
+
+        </>
+    )
+}
+export default ShowProductEdit
+
